LayoutTests/media/track/video-track-configuration-mp4-default-colorspace-expected.txt


RUN(video.src = "../content/video-as-img.mp4")
EVENT(canplay)
EXPECTED (video.videoTracks.length == '1') OK
EXPECTED (video.videoTracks[0].configuration.colorSpace.matrix != 'null') OK
EXPECTED (video.videoTracks[0].configuration.colorSpace.primaries != 'null') OK
EXPECTED (video.videoTracks[0].configuration.colorSpace.transfer != 'null') OK
END OF TEST
